+++ main.go
@@ -13,7 +13,7 @@ var (
 )
 
 func main() {
-	if err := cmd.ExecuteMinimal(); err != nil {
+	if err := cmd.Execute(); err != nil {
 		fmt.Fprintf(os.Stderr, "Error: %v\n", err)
 		os.Exit(1)
 	}
